Move VIES helper to shared package
