+++ backend/app/api/src/endpoints/organization/dashboard/invoices/PatchInvoicesEndpoint.ts
@@ -8,7 +8,7 @@ import { AuthenticatedStructures } from '../../../../helpers/AuthenticatedStruct
 import { Context } from '../../../../helpers/Context.js';
 import { Invoice } from '@stamhoofd/models';
 import { SimpleError } from '@simonbackx/simple-errors';
-import { ViesHelper } from '../../../../helpers/ViesHelper.js';
+import { ViesHelper } from '@stamhoofd/vies/ViesHelper.js';
 import { InvoiceService } from '../../../../services/InvoiceService.js';
 
 type Params = Record<string, never>;
